Table Modifications
Because the values for video games are currently hard-coded, it's easy to put them in a table. Originally I created a list of Members, which doesn't really make sense because each Member has more than one attribute. The next step is to move both of these Models(Member and Video Game) into tables.